Cambios estéticos y cambio de nombres
Cambio de nombre de formulario principal(form1 a mainForm)
Cambio de título de formulario principal(form 1 a Gestión de inventario)
Cambio de icono
Creación carpeta img para guardar recursos visuales (como el icono añadido)
