--- Prueba_proyecto final/Prueba_proyecto final.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36414.22 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Prueba_proyecto final", "Prueba_proyecto final\Prueba_proyecto final.csproj", "{92552607-385E-47F9-8A02-D87A96EB9420}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{92552607-385E-47F9-8A02-D87A96EB9420}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{92552607-385E-47F9-8A02-D87A96EB9420}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{92552607-385E-47F9-8A02-D87A96EB9420}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{92552607-385E-47F9-8A02-D87A96EB9420}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {A4BE3D2D-6928-4597-9C4C-826AB5AAF94C}
-	EndGlobalSection
-EndGlobal
